\name{block.info.per.su.f}
\alias{block.info.per.su.f}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Sorts block information for each sampling unit }
\description{
    Organizes block information for each sampling unit.
}
\usage{
block.info.per.su.f(block = block.size, data = dsm.data, name.su = unique.sampling.units)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{block}{ Size of block (specified by user) }
  \item{data}{ Original response variable by segment }
  \item{name.su}{ Number of segment for entire data set }
}
%\details{
%}
%\value{
%}
%\references{ ~put references to the literature/web site here ~ }
\author{ Louise Burt, RUWPA, \email{louise@mcs.st-and.ac.uk} }
%\note{ ~~further notes~~ 
%}
%\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{  }
