% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.info.per.su.R
\name{block.info.per.su}
\alias{block.info.per.su}
\title{Find the block information}
\usage{
block.info.per.su(block.size, data, name.su)
}
\arguments{
\item{block.size}{number of segments per block}

\item{data}{data used to build the model}

\item{name.su}{names of the sampling units (i.e., transects)}
}
\value{
a \code{data.frame} with the following columns
\itemize{
\item \code{name}        the sample unit name (e.g. transect label)
\item \code{num.seg}     number of segments in that transect
\item \code{num.block}   number of blocks available
\item \code{start.block} block number for first block
\item \code{end.block}   block number for last block
\item \code{num.req}     number of blocks needed for the unit
}
}
\description{
Takes the transect data and works out how many blocks of a given
size (in segment terms) fit into each.
}
