% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.cols.R
\name{check.cols}
\alias{check.cols}
\title{Check column names exist}
\usage{
check.cols(ddf.obj, segment.data, observation.data, segment.area)
}
\arguments{
\item{ddf.obj}{a \code{ddf} object from \code{\link{mrds}}}

\item{segment.data}{segment data as defined in \code{\link{dsm}}}

\item{observation.data}{observation data as defined in \code{\link{dsm}}}

\item{segment.area}{area of segments}
}
\value{
nothing, but throws an error if something went wrong
}
\description{
Internal function to check that supplied \code{data.frames} have the correct
columns and checks that sample labels are all unique.
}
\author{
David Lawrence Miller
}
