% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_Ke_Cl}
\alias{df_Ke_Cl}
\title{Kenya Environmental covariates and observations}
\format{
An object of class \code{data.frame} with 400 rows and 4 columns.
}
\source{
<https://doi.org/10.1016/j.geoderma.2009.04.024>
}
\usage{
df_Ke_Cl
}
\description{
df_Ke_Cl : the soil clay observations of Kenya
Ke_covariates: the environmental covariates
}
\keyword{datasets}
