% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dead_functions.R
\name{dsm.var.gam}
\alias{dsm.var.gam}
\title{Prediction variance estimation assuming independence}
\usage{
dsm.var.gam(
  dsm.obj,
  pred.data,
  off.set,
  seglen.varname = "Effort",
  type.pred = "response"
)
}
\arguments{
\item{dsm.obj}{a model object fitted by \code{\link{dsm}}.}

\item{pred.data}{either: a single prediction grid or list of prediction
grids. Each grid should be a \code{data.frame} with the same columns as the
original data.}

\item{off.set}{a a vector or list of vectors with as many elements as there
are in \code{pred.data}. Each vector is as long as the number of rows in the
corresponding element of \code{pred.data}. These give the area associated with
each prediction cell. If a single number is supplied it will be replicated
for the length of \code{pred.data}.}

\item{seglen.varname}{name for the column which holds the segment length
(default value \code{"Effort"}).}

\item{type.pred}{should the predictions be on the "response" or "link"
scale? (default \code{"response"}).}
}
\description{
This function is deprecated, use \link{dsm_var_gam}.
}
