% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_ddf.R
\name{dummy_ddf}
\alias{dummy_ddf}
\title{Detection function objects when detection is certain}
\usage{
dummy_ddf(object, size = 1, width, left = 0, transect = "line")
}
\arguments{
\item{object}{numeric vector of object identifiers, relating to the \code{object} field in the observation data of the DSM.}

\item{size}{group size for each observation (default all groups size 1)}

\item{width}{right truncation}

\item{left}{left truncation (default 0, no left truncation)}

\item{transect}{\code{"line"} or \code{"point"} transect}
}
\description{
Create a detection function object for strip/plot surveys for use in density surface models.
}
\author{
David L Miller
}
