\name{get.model.details.f}
\alias{get.model.details.f}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Parse specifics of fitted density surface model }
\description{
  For performing the parametric bootstrap, details of the fitted model are required, and these are derived courtesy of this function
}
\usage{
get.model.details.f(modelname = modelname)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{modelname}{ Density surface model that was fitted to data }
}
%\details{
%}
\value{
  List that describes the details of the model being fitted
}
%\references{ ~put references to the literature/web site here ~ }
\author{ Louise Burt, RUWPA, \email{louise@mcs.st-and.ac.uk} }
%\note{ ~~further notes~~ 
%}
%\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{  }

