% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_cov_overlay.R
\name{obs_cov_overlay}
\alias{obs_cov_overlay}
\title{Overlay between observations and  environmental covariates}
\usage{
obs_cov_overlay(df, covariates)
}
\arguments{
\item{df}{Dataframe containing the observation, sampling points}

\item{covariates}{Environmental covariates for the area of interest}
}
\value{
A Dataframe with including your observations and environmental covariates at locations samples locations
}
\description{
Overlay and intersection between observations and  environmental covariates in the area of interest. It finds the covariates values at the locations of samples/Observations
}
\examples{
##Load the environmental covariates in area
covariates <- load("Covariates")
##Load the observations
df <- load("Observations")
m <- obs_cov_overlay(df,covariates)

#OR use the data in this package as observations and covariates; the data related to Kenya
data(df_Ke_Cl)
data(Ke_covariates)
Ke_Cl <- obs_cov_overlay(df_Ke_Cl,Ke_covariates)

}
\author{
Fatemeh Hateffard and Luc Steinbuch
}
