% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dsm.R
\name{plot.dsm}
\alias{plot.dsm}
\title{Plot a density surface model.}
\usage{
\method{plot}{dsm}(x, ...)
}
\arguments{
\item{x}{a model fitted by \code{\link{dsm}}}

\item{\dots}{other arguments passed to \code{\link[mgcv:plot.gam]{plot.gam}}.}
}
\value{
a plot!
}
\description{
See \code{\link[mgcv:plot.gam]{plot.gam}}.
}
\seealso{
\code{\link{dsm}} \code{\link[mgcv:plot.gam]{plot.gam}}
}
\author{
David L. Miller
}
