% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_ddf.R
\name{predict.fake_ddf}
\alias{predict.fake_ddf}
\title{Prediction for fake detection functions}
\usage{
\method{predict}{fake_ddf}(object, newdata = NULL, compute = FALSE,
  int.range = NULL, esw = FALSE, ...)
}
\arguments{
\item{object}{model object}

\item{newdata}{how many 1s should we return?}
}
\description{
Dummy function to return the correct number of 1s.
}
