% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark_ddf.R
\name{predict.rmark_ddf}
\alias{predict.rmark_ddf}
\title{Prediction for RMark-dervied "detection" functions}
\usage{
\method{predict}{rmark_ddf}(
  object,
  newdata = NULL,
  compute = FALSE,
  int.range = NULL,
  esw = FALSE,
  ...
)
}
\arguments{
\item{object}{model object}

\item{newdata}{covariates needed to get a probabilities of detection}

\item{compute}{unused, compatability with \code{\link[mrds:predict.ds]{mrds::predict}}}

\item{int.range}{unused, compatability with \code{\link[mrds:predict.ds]{mrds::predict}}}

\item{esw}{unused, compatability with \code{\link[mrds:predict.ds]{mrds::predict}}}

\item{\dots}{for S3 consistency}
}
\description{
Prediction function for detection functions dervied from RMark models.
}
\author{
David L Miller
}
