\name{print.dsm}
\alias{print.dsm}
\title{Print a description of a density surface model object}
\usage{
\method{print}{dsm}(x, ...)
}
\arguments{
  \item{x}{a \code{dsm} object}

  \item{\dots}{unspecified and unused arguments for S3
  consistency}
}
\value{
NULL
}
\description{
This method just gives a short description of the fitted
model. Use the \code{\link{summary.dsm}} method for more
information.
}
\author{
David L. Miller
}
\seealso{
\code{\link{summary.ds}}
}
\keyword{utility}

