% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dsm.var.R, R/print.dsm_varprop.R
\name{print.dsm.var}
\alias{print.dsm.var}
\title{Print a description of a density surface model variance object}
\usage{
\method{print}{dsm.var}(x, ...)

\method{print}{dsm.var}(x, ...)
}
\arguments{
\item{x}{a \code{dsm} variance object}

\item{\dots}{unspecified and unused arguments for S3 consistency}

\item{x}{a \code{dsm} variance object}

\item{\dots}{unspecified and unused arguments for S3 consistency}
}
\description{
This method only provides a short summary, use the 
\code{\link{summary.dsm.var}} method for information.

This method only provides a short summary, see \code{\link{summary.dsm_varprop}}.
}
\author{
David L. Miller

David L. Miller
}
\seealso{
\code{\link{summary.dsm.var}}

\code{\link{summary.dsm_varprop}}
}
\keyword{utility}

