% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dsm_varprop.R
\name{print.dsm_varprop}
\alias{print.dsm_varprop}
\title{Print a description of a density surface model variance object}
\usage{
\method{print}{dsm_varprop}(x, ...)
}
\arguments{
\item{x}{a \code{dsm} variance object}

\item{\dots}{unspecified and unused arguments for S3 consistency}
}
\description{
This method only provides a short summary, see \code{\link{summary.dsm_varprop}}.
}
\seealso{
\code{\link{summary.dsm_varprop}}
}
\author{
David L. Miller
}
\keyword{utility}
