\name{print.summary.dsm}
\alias{print.summary.dsm}
\title{Print summary of density surface model object}
\usage{
  \method{print}{summary.dsm} (x, ...)
}
\arguments{
  \item{x}{a summary of \code{dsm} object}

  \item{\dots}{unspecified and unused arguments for S3
  consistency}
}
\value{
  NULL
}
\description{
  See \code{\link{summary.dsm}} for information.
}
\author{
  David L. Miller
}
\seealso{
  \code{\link{summary.ds}}
}
\keyword{utility}

