\name{print.summary.dsm.var}
\alias{print.summary.dsm.var}
\title{Print summary of density surface model variance object}
\usage{
  \method{print}{summary.dsm.var} (x, ...)
}
\arguments{
  \item{x}{a summary of \code{dsm} variance object}

  \item{\dots}{unspecified and unused arguments for S3
  consistency}
}
\value{
  NULL
}
\description{
  See \code{\link{summary.dsm.var}} for information.
}
\author{
  David L. Miller
}
\seealso{
  \code{\link{summary.ds.var}}
}
\keyword{utility}

