% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.dsm_varprop.R
\name{print.summary.dsm_varprop}
\alias{print.summary.dsm_varprop}
\title{Print summary of density surface model variance object}
\usage{
\method{print}{summary.dsm_varprop}(x, ...)
}
\arguments{
\item{x}{a summary of \code{dsm} variance object}

\item{\dots}{unspecified and unused arguments for S3 consistency}
}
\value{
\code{NULL}
}
\description{
See \code{\link{summary.dsm_varprop}} for information.
}
\seealso{
\code{\link{summary.dsm.var}}
}
\author{
David L. Miller
}
\keyword{utility}
