\name{resids.when.missing}
\alias{resids.when.missing}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Computes residuals from predicted density surface }
\description{
  Compute log of residuals from predicted density surface taking into account possible missing values
}
\usage{
resids.when.missing(model = model.object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{ Object containing result of model fit }
}
%\details{
%}
\value{
  Residuals to be shuffled about in subsequent parametric bootstrapping
}
%\references{ ~put references to the literature/web site here ~ }
\author{ Louise Burt, RUWPA, \email{louise@mcs.st-and.ac.uk} }
%\note{ ~~further notes~~ 
%}
%\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{  }

