% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFmodel.R
\name{RFmodel}
\alias{RFmodel}
\title{Random Forest model}
\usage{
RFmodel(df, covariates, variable_of_interest)
}
\arguments{
\item{df}{A Dataframe including observations and environmental covariates at samples locations}

\item{covariates}{Environmental covariates for the area of interest}

\item{variable_of_interest}{Name or abbreviation of the target variable available in "df" dataframe}
}
\value{
A Random Forest model with 10 fold cross-validation
}
\description{
Training a random forest model, use 10-fold cross validation without optimization
}
\examples{
#in this example, Clay content in soil is intended to predict by random forest model
Clay_model<- RFmodel(df,covariates,variable_of_interest="Cl")
}
\author{
Fatemeh Hateffard and Luc Steinbuch
}
