% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark_ddf.R
\name{rmark_ddf}
\alias{rmark_ddf}
\title{Using RMark models in density surface models}
\usage{
rmark_ddf(rmark_model, obs, width, left = 0, transect = "line")
}
\arguments{
\item{rmark_model}{model fitted by RMark}

\item{obs}{observation data, \code{data.frame} with columns \code{object} and \code{size}
plus all covariates used in \code{rmark_model}.}

\item{width}{right truncation}

\item{left}{left truncation (default 0, no left truncation)}

\item{transect}{\code{"line"} or \code{"point"} transect}
}
\description{
Create a detection function or availability object from a fitted RMark model.
}
\details{
Models fitted by RMark can be used in DSM.... CHAT HERE

RMark models can't contain variables named \code{detected}, \code{distance} or \code{observer}.
\code{Time} is the observer id

Obs should only contain \emph{one} observation per detected group!
}
\author{
David L Miller
}
