\name{summary.dsm.var}
\alias{summary.dsm.var}
\title{Summarize the varianc of a density surface model}
\usage{
  \method{summary}{dsm.var} (object, alpha = 0.05, ...)
}
\arguments{
  \item{object}{a \code{dsm.var} object}

  \item{alpha}{alpha level for confidence intervals}

  \item{\dots}{unused arguments for S3 compatibility}
}
\value{
  a summary object
}
\description{
  Gives a brief summary of a fitted \code{dsm} variance
  object.
}
\author{
  David L. Miller
}

