% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dsm_varprop.R
\name{summary.dsm_varprop}
\alias{summary.dsm_varprop}
\title{Summarize the variance of a density surface model}
\usage{
\method{summary}{dsm_varprop}(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{a \code{dsm.var} object}

\item{alpha}{alpha level for confidence intervals (default 0.05 to give a 95\% confidence internal)}

\item{\dots}{unused arguments for S3 compatibility}
}
\value{
a summary object
}
\description{
Gives a brief summary of a fitted \code{dsm_varprop} variance object.
}
\seealso{
dsm_varprop summary.dsm.var
}
\author{
David L. Miller
}
