% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{transpose}
\alias{transpose}
\title{Data Set Manipulation}
\usage{
transpose(data)
}
\arguments{
\item{data}{ an n \eqn{x} m data frame}
}
\description{
This function allows transpose an data set.
}
\author{
Luiz Henrique Leal \email{lhleal@inmetro.gov.br}
}
\seealso{
  \code{\link{unstack2}}.
}
\examples{
 dfr=infert[1:6,1:2]
 unstack2(dfr)
 transpose(unstack2(dfr))
}
