% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{unstack2}
\alias{unstack2}
\title{Data Set Manipulation}
\usage{
unstack2(data)
}
\arguments{
\item{data}{ an n \eqn{x} 2 data frame}
}
\description{
This function allows unstack an data set.
}
\author{
Luiz Henrique Leal \email{lhleal@inmetro.gov.br}
}
\seealso{
  \code{\link{transpose}}.
}
\examples{
 dfr=infert[1:6,1:2] ; dfr
 unstack2(dfr)
}
