% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsm-package.R
\name{dsm-package}
\alias{dsm-package}
\title{Density surface modelling}
\description{
\code{dsm} implements spatial models for distance sampling data.
}
\details{
Further information on distance sampling methods and example code is available at \url{http://distancesampling.org/R/}.

For help with distance sampling and this package, there is a Google Group \url{https://groups.google.com/forum/#!forum/distance-sampling}.
}

