% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm.var.movblk.R
\name{dsm.var.movblk}
\alias{dsm.var.movblk}
\title{Variance estimation via parametric moving block bootstrap}
\usage{
dsm.var.movblk(dsm.object, pred.data, n.boot, block.size, off.set,
  ds.uncertainty = FALSE, samp.unit.name = "Transect.Label",
  progress.file = NULL, bs.file = NULL, bar = TRUE)
}
\arguments{
\item{dsm.object}{object returned from \code{\link{dsm}}.}

\item{pred.data}{either: a single prediction grid or list of prediction grids. Each grid should be a \code{data.frame} with the same columns as the original data.}

\item{n.boot}{number of bootstrap resamples.}

\item{block.size}{number of segments in each block.}

\item{off.set}{a a vector or list of vectors with as many elements as there are in \code{pred.data}. Each vector is as long as the number of rows in the corresponding element of \code{pred.data}. These give the area associated with each prediction cell. If a single number is supplied it will be replicated for the length of \code{pred.data}.}

\item{ds.uncertainty}{incorporate uncertainty in the detection function? See
Details, below. Note that this feature is EXPERIMENTAL at the moment.}

\item{samp.unit.name}{name sampling unit to resample (default
'Transect.Label').}

\item{progress.file}{path to a file to be used (usually by Distance) to
generate a progress bar (default \code{NULL} -- no file written).}

\item{bs.file}{path to a file to store each boostrap round. This stores all
of the bootstrap results rather than just the summaries, enabling
outliers to be detected and removed. (Default \code{NULL}).}

\item{bar}{should a progress bar be printed to screen? (Default \code{TRUE}).}
}
\description{
Estimate the variance in abundance over an area using a moving block
bootstrap. Two procedures are implemented, one incorporating detection
function uncertainty, one not.
}
\section{Details}{

 Setting \code{ds.uncertainty=TRUE} will incorporate detection function
 uncertainty directly into the bootstrap. This is done by generating
 observations from the fitted detection function and then re-fitting a new
 detection function (of the same form), then calculating a new effective
 strip width. Rejection sampling is used to generate the observations (except
 in the half-normal case) so the procedure can be rather slow. Note that this
 is currently not supported with covariates in the detection function.

 Setting \code{ds.uncertainty=FALSE} will incorporate detection function
 uncertainty using the delta method. This assumes that the detection function
 and the spatial model are INDEPENDENT. This is probably not reasonable.
}
\examples{
\dontrun{
library(Distance)
library(dsm)

# load the Gulf of Mexico dolphin data (see ?mexdolphins)
data(mexdolphins)
attach(mexdolphins)

# fit a detection function and look at the summary
hr.model <- ds(distdata, max(distdata$distance),
               key = "hr", adjustment = NULL)
summary(hr.model)

# fit a simple smooth of x and y
mod1<-dsm(N~s(x,y), hr.model, segdata, obsdata)
summary(mod1)

# calculate the variance by 500 moving block bootstraps
mod1.movblk <- dsm.var.movblk(mod1, preddata, n.boot = 500,
   block.size = 3, samp.unit.name = "Transect.Label",
   off.set = preddata$area,
   bar = TRUE, bs.file = "mexico-bs.csv", ds.uncertainty = TRUE)
# detach the data
detach("mexdolphins")
}
}

