% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_ddf.R
\name{dummy_ddf}
\alias{dummy_ddf}
\title{Make a dummy detection function for strip transects}
\usage{
dummy_ddf(obs, size = 1, width, left = 0, transect = "line")
}
\arguments{
\item{obs}{which observations are to be included? Numeric vector corresponding to object IDs}

\item{size}{group size of observed group (default all groups size 1)}

\item{width}{right truncation}

\item{left}{left truncation}

\item{transect}{"line" or "point" transect (character)}
}
\description{
Strip transects aren't really detection functions, but this allows you to use them like detection functions in \code{dsm}.
}
