\name{plot.dsm}
\alias{plot.dsm}
\title{Plot a density surface model.}
\usage{
  \method{plot}{dsm} (x, ...)
}
\arguments{
  \item{x}{a \code{dsm} object}

  \item{\dots}{other arguments passed to
  \code{\link{plot.gam}}.}
}
\value{
  a plot!
}
\description{
  See \code{\link{plot.gam}}.
}
\author{
  David L. Miller
}
\seealso{
  dsm plot.gam
}

