% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmark_ddf.R
\name{predict.rmark_ddf}
\alias{predict.rmark_ddf}
\title{Prediction for fake detection functions}
\usage{
\method{predict}{rmark_ddf}(
  object,
  newdata = NULL,
  compute = FALSE,
  int.range = NULL,
  esw = FALSE,
  ...
)
}
\arguments{
\item{object}{model object}

\item{newdata}{covariates needed to get a probabilities of detection}

\item{compute}{unused, compatability with \code{\link[mrds:predict.ds]{mrds::predict}}}

\item{int.range}{unused, compatability with \code{\link[mrds:predict.ds]{mrds::predict}}}

\item{esw}{should the strip width be returned?}

\item{\dots}{for S3 consistency}
}
\description{
Prediction function for dummy detection functions. The function returns as
many 1s as there are rows in \code{newdata}. If \code{esw=TRUE} then the
strip width is returned.
}
\author{
David L Miller
}
