% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dsm.R
\name{print.dsm}
\alias{print.dsm}
\title{Print a description of a density surface model object}
\usage{
\method{print}{dsm}(x, ...)
}
\arguments{
\item{x}{a model fitted by \code{\link{dsm}}}

\item{\dots}{unspecified and unused arguments for S3 consistency}
}
\value{
\code{NULL}
}
\description{
This method just gives a short description of the fitted model. Use the
\code{\link{summary.dsm}} method for more information.
}
\author{
David L. Miller
}
\keyword{utility}
