% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.dsm.var}
\alias{print.dsm.var}
\title{Print a description of a density surface model variance object}
\usage{
\method{print}{dsm.var}(x, ...)
}
\arguments{
\item{x}{a \code{dsm} variance object}

\item{\dots}{unspecified and unused arguments for S3 consistency}
}
\description{
This method only provides a short summary, use the
\code{\link{summary.dsm.var}} method for information.
}
\author{
David L. Miller
}
\seealso{
\code{\link{summary.dsm.var}}
}
\keyword{utility}

