% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.summary.dsm.var}
\alias{print.summary.dsm.var}
\title{Print summary of density surface model variance object}
\usage{
\method{print}{summary.dsm.var}(x, ...)
}
\arguments{
\item{x}{a summary of \code{dsm} variance object}

\item{\dots}{unspecified and unused arguments for S3 consistency}
}
\description{
See \code{\link{summary.dsm.var}} for information.
}
\author{
David L. Miller
}
\seealso{
\code{\link{summary.dsm.var}}
}
\keyword{utility}

