% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dead_functions.R
\name{rqgam.check}
\alias{rqgam.check}
\title{Randomised quantile residuals check plot for GAMs/DSMs}
\usage{
rqgam.check(gam.obj, ...)
}
\arguments{
\item{gam.obj}{a \code{\link[mgcv:gam]{gam}}, \code{\link[stats:glm]{glm}} or \code{\link{dsm}}
object.}

\item{...}{arguments passed on to all plotting functions}
}
\description{
This function is deprecated, use \link{rqgam_check}.
}
