% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dsm.R
\name{summary.dsm}
\alias{summary.dsm}
\title{Summarize a fitted density surface model}
\usage{
\method{summary}{dsm}(object, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{dsm}}}

\item{\dots}{other arguments passed to \code{\link[mgcv:summary.gam]{summary.gam}}}
}
\value{
a summary object
}
\description{
Gives a brief summary of a fitted \code{\link{dsm}} object.
}
\seealso{
\code{\link{dsm}}
}
\author{
David L. Miller
}
