% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.dsm}
\alias{summary.dsm}
\title{Summarize a fitted density surface model}
\usage{
\method{summary}{dsm}(object, ...)
}
\arguments{
\item{object}{a \code{dsm} object}

\item{\dots}{other arguments passed to \code{\link{summary.gam}}.}
}
\value{
a summary object
}
\description{
Gives a brief summary of a fitted \code{dsm} object.
}
\author{
David L. Miller
}
\seealso{
dsm
}

