% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dead_functions.R
\name{vis.concurvity}
\alias{vis.concurvity}
\title{Visualise concurvity between terms in a GAM}
\usage{
vis.concurvity(model, type = "estimate")
}
\arguments{
\item{model}{fitted model}

\item{type}{concurvity measure to plot, see \code{\link[mgcv:concurvity]{concurvity}}}
}
\description{
This function is deprecated, use \link{vis_concurvity}.
}
